Restructure the main AST pass loop.
Instead of running passes from all plugins interleaved at the same time,
we now run plugins subsequently. For now, this works only for the HILTI
plugin, the Spicy plugin is broken until we rework how that's
implemented.

This also replaces the three different validation hooks with just a
single, in anticipation that in the future that's all we'll need.

All HILTI tests pass, except for one using spicy-build.